fix: refactor CanaryPopup-related files
